// types/Card.ts
export type Card = {
    id: string;
    badges: any;
    checkItemStates: any[];
    closed: boolean;
    dueComplete: boolean;
    dateLastActivity: string;
    desc: string;
    descData: any;
    due: string | null;
    dueReminder: number | null;
    email: string | null;
    idBoard: string;
    idChecklists: string[];
    idList: string;
    idMembers: string[];
    idMembersVoted: string[];
    idShort: number;
    idAttachmentCover: string | null;
    labels: any[];
    idLabels: string[];
    manualCoverAttachment: boolean;
    name: string;
    nodeId: string;
    pinned: boolean;
    pos: number;
    shortLink: string;
    shortUrl: string;
    start: string | null;
    subscribed: boolean;
    url: string;
    cover: any;
    isTemplate: boolean;
    cardRole: string | null;
    mirrorSourceId: string | null;
};